review 1

+++ foodgram/recipes/forms.py
@@ -1,49 +1,51 @@
 from django import forms
+from django.core.exceptions import ObjectDoesNotExist
 
-from .models import Recipe, AmountOfIngredients, Ingredient
+from recipes.models import AmountOfIngredient
+from recipes.models import Ingredient
+from recipes.models import Recipe
 
 
 class RecipeForm(forms.ModelForm):
 
     def __init__(self, data=None, *args, **kwargs):
+        self.ingredients = {}
         if data is not None:
             data = data.copy()
-            self.ingredients = self.get_ingredients(data)
+            self.get_ingredients(data)
 
         super().__init__(data=data, *args, **kwargs)
 
-    @staticmethod
-    def get_ingredients(data):
-        ingredients = {}
+    def get_ingredients(self, data):
         for key, name in data.items():
             if key.startswith('nameIngredient'):
-                number = key.split('_')[1]
+                _, _, number = key.partition('_')
                 value = f'valueIngredient_{number}'
-                ingredients[name] = data.get(value)
-
-        return ingredients
+                self.ingredients[name] = {'amount': int(data.get(value))}
 
     def save(self, commit=True):
         recipe = super().save(commit=False)
         recipe.save()
         objects = []
-        for title, amount in self.ingredients.items():
-            ingredient = Ingredient.objects.filter(title=title)
-            if ingredient.exists():
-                objects.append(AmountOfIngredients(recipe=recipe,
-                                                   ingredient=ingredient.get(),
-                                                   amount=int(amount), )
-                               )
-
-        recipe.amount.all().delete()
-        AmountOfIngredients.objects.bulk_create(objects)
+        for data in self.ingredients.values():
+            objects.append(AmountOfIngredient(recipe=recipe,
+                                              ingredient=data.get('object'),
+                                              amount=data.get('amount'), )
+                           )
+        if objects:
+            recipe.amount.all().delete()
+            AmountOfIngredient.objects.bulk_create(objects)
         self.save_m2m()
         return recipe
 
     def clean(self):
+        if not self.ingredients:
+            raise forms.ValidationError('Empty ingredients list not allowed')
         for title, amount in self.ingredients.items():
-            ingredient = Ingredient.objects.filter(title=title)
-            if not ingredient.exists():
+            try:
+                ingredient = Ingredient.objects.filter(title=title).get()
+                self.ingredients[title].update({'object': ingredient})
+            except ObjectDoesNotExist:
                 raise forms.ValidationError(f'{title} not valid ingredient')
         return super().clean()
 
@@ -53,5 +55,5 @@ class RecipeForm(forms.ModelForm):
                   'image',
                   'description',
                   'tags',
-                  'cook_time', )
+                  'cook_time',)
         widgets = {'tags': forms.CheckboxSelectMultiple(), }
